PR Feedback: Break out Protocols to own files
