+++ Sources/Subscription/Managers/StorePurchaseManager/StorePurchaseManager.swift
@@ -34,144 +34,6 @@ public enum StorePurchaseManagerError: Error {
     case unknownError
 }
 
-/// A protocol that defines the properties of an introductory offer for a subscription product.
-/// Use this protocol to represent trial periods, introductory prices, or other special offers.
-@available(macOS 12.0, iOS 15.0, *)
-public protocol SubscriptionProductIntroductoryOffer {
-    /// The unique identifier of the introductory offer.
-    var id: String? { get }
-
-    /// The formatted price of the offer that should be displayed to users.
-    var displayPrice: String { get }
-
-    /// The duration of the offer in days.
-    var periodInDays: Int { get }
-
-    /// Indicates whether this offer represents a free trial period.
-    var isFreeTrial: Bool { get }
-}
-
-/// Extends StoreKit's Product.SubscriptionOffer to conform to SubscriptionProductIntroductoryOffer.
-@available(macOS 12.0, iOS 15.0, *)
-extension Product.SubscriptionOffer: SubscriptionProductIntroductoryOffer {
-    /// Calculates the total number of days in the offer period by multiplying
-    /// the base period length by the period count.
-    public var periodInDays: Int {
-        period.periodInDays * periodCount
-    }
-
-    /// Determines if this offer represents a free trial based on the payment mode.
-    public var isFreeTrial: Bool {
-        paymentMode == .freeTrial
-    }
-}
-
-/// A protocol that defines the core functionality and properties of a subscription product.
-/// Conforming types must provide information about pricing, description, and subscription terms.
-@available(macOS 12.0, iOS 15.0, *)
-public protocol SubscriptionProduct: Equatable {
-    /// The unique identifier of the product.
-    var id: String { get }
-
-    /// The user-facing name of the product.
-    var displayName: String { get }
-
-    /// The formatted price that should be displayed to users.
-    var displayPrice: String { get }
-
-    /// A detailed description of the product.
-    var description: String { get }
-
-    /// Indicates whether this is a monthly subscription.
-    var isMonthly: Bool { get }
-
-    /// Indicates whether this is a yearly subscription.
-    var isYearly: Bool { get }
-
-    /// The introductory offer associated with this subscription, if any.
-    var introductoryOffer: SubscriptionProductIntroductoryOffer? { get }
-
-    /// Indicates whether this subscription has a Free Trial offer available.
-    var hasFreeTrialOffer: Bool { get }
-
-    /// Asynchronously determines whether the user is eligible for an introductory offer.
-    var isEligibleForIntroOffer: Bool { get async }
-
-    /// Initiates a purchase of the subscription with the specified options.
-    /// - Parameter options: A set of options to configure the purchase.
-    /// - Returns: The result of the purchase attempt.
-    /// - Throws: An error if the purchase fails.
-    func purchase(options: Set<Product.PurchaseOption>) async throws -> Product.PurchaseResult
-}
-
-/// Extends StoreKit's Product to conform to SubscriptionProduct.
-@available(macOS 12.0, iOS 15.0, *)
-extension Product: SubscriptionProduct {
-    /// Determines if this is a monthly subscription by checking if the subscription period
-    /// is exactly one month.
-    public var isMonthly: Bool {
-        guard let subscription else { return false }
-        return subscription.subscriptionPeriod.unit == .month &&
-        subscription.subscriptionPeriod.value == 1
-    }
-
-    /// Determines if this is a yearly subscription by checking if the subscription period
-    /// is exactly one year.
-    public var isYearly: Bool {
-        guard let subscription else { return false }
-        return subscription.subscriptionPeriod.unit == .year &&
-        subscription.subscriptionPeriod.value == 1
-    }
-
-    /// Returns the introductory offer for this subscription if available.
-    public var introductoryOffer: (any SubscriptionProductIntroductoryOffer)? {
-        subscription?.introductoryOffer
-    }
-
-    /// Indicates whether this subscription has a Free Trial offer.
-    public var hasFreeTrialOffer: Bool {
-        return subscription?.introductoryOffer?.isFreeTrial ?? false
-    }
-
-    /// Asynchronously checks if the user is eligible for an introductory offer.
-    public var isEligibleForIntroOffer: Bool {
-        get async {
-            guard let subscription else { return false }
-            return await subscription.isEligibleForIntroOffer
-        }
-    }
-
-    /// Implements Equatable by comparing product IDs.
-    public static func == (lhs: Product, rhs: Product) -> Bool {
-        return lhs.id == rhs.id
-    }
-}
-
-/// A protocol for types that can fetch subscription products.
-@available(macOS 12.0, iOS 15.0, *)
-public protocol ProductFetching {
-    /// Fetches products for the specified identifiers.
-    /// - Parameter identifiers: An array of product identifiers to fetch.
-    /// - Returns: An array of subscription products.
-    /// - Throws: An error if the fetch operation fails.
-    func products(for identifiers: [String]) async throws -> [any SubscriptionProduct]
-}
-
-/// A default implementation of ProductFetching that uses StoreKit's standard product fetching.
-@available(macOS 12.0, iOS 15.0, *)
-public final class DefaultProductFetcher: ProductFetching {
-    /// Initializes a new DefaultProductFetcher instance.
-    public init() {}
-
-    /// Fetches products using StoreKit's Product.products API.
-    /// - Parameter identifiers: An array of product identifiers to fetch.
-    /// - Returns: An array of subscription products.
-    /// - Throws: An error if the fetch operation fails.
-    public func products(for identifiers: [String]) async throws -> [any SubscriptionProduct] {
-        return try await Product.products(for: identifiers)
-    }
-}
-
 public protocol StorePurchaseManager {
     typealias TransactionJWS = String
 
@@ -510,21 +372,6 @@ private extension SubscriptionOption {
         }
 
         self.init(id: product.id, cost: .init(displayPrice: product.displayPrice, recurrence: recurrence), offer: offer)
-    }
-}
-
-@available(macOS 12.0, iOS 15.0, *)
-private extension Product.SubscriptionPeriod {
-
-    var periodInDays: Int {
-        switch unit {
-        case .day: return value
-        case .week: return value * 7
-        case .month: return value * 30
-        case .year: return value * 365
-        @unknown default:
-            return value
-        }
     }
 }
